Split up the header files so the classes have their own dedicated header files
