+++ Program 2 for Programming Systems/algos.cpp
@@ -1,5 +1,6 @@
 #include "node.h"
-
+#include "questions.h"
+#include "algos.h"
 
 algos::algos(){
     bigO = NULL;
