+++ Program 2 for Programming Systems/list.cpp
@@ -1,5 +1,5 @@
 #include "table.h"
-
+#include "questions.h"
 
 list::list(){
