+++ Program 2 for Programming Systems/node.cpp
@@ -1,4 +1,8 @@
 #include "node.h"
+#include "questions.h"
+#include "algos.h"
+#include "threads.h"
+#include "parallel.h"
 
 //node constructor to set data
 node::node(){
