+++ Program 2 for Programming Systems/parallel.cpp
@@ -1,5 +1,6 @@
 #include "node.h"
-
+#include "questions.h"
+#include "parallel.h"
 
 parallel::parallel(){
     cores = 0;
