+++ Program 2 for Programming Systems/table.h
@@ -4,6 +4,7 @@
  */
 
 #include "node.h"
+#include "list.h"
 
 const int NUM = 3;
 
@@ -31,27 +32,5 @@ class table
         //int dis; //int that allows user to choose which type of question to display
 };
 
-//the list class manages the insertion of nodes
-class list
-{
-    public:
-        list();
-        ~list();
-      //  void insertw(tabl); //wrapper for insert
-        void display(); //wrapper for display funct
-        void remove(); //remove question
-        void insert(int);
-        void search(int ref); //search function 
-        void test();
-         int delete_item(int del);
-
-    protected:
-        int delete_item(node*& head, int del);
-         node* head; //contains a node for insertion
-        void display(node* head);
-        void remove(node*& head);
-       // node* search(node* head); //search function 
-        int traverse(node* head, int ref);
-};
