+++ Program 2 for Programming Systems/threads.cpp
@@ -1,5 +1,6 @@
 #include "node.h"
-
+#include "questions.h"
+#include "threads.h"
 
 threads::threads(){
     tc = 0;
